Update testing to support new syntax
