+++ src/test/java/seedu/vms/logic/parser/BasicParserTest.java
@@ -1,0 +1,25 @@
+package seedu.vms.logic.parser;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import org.junit.jupiter.api.Test;
+
+import seedu.vms.logic.commands.ExitCommand;
+import seedu.vms.logic.commands.HelpCommand;
+
+public class BasicParserTest {
+    private final BasicParser parser = new BasicParser();
+
+
+    @Test
+    public void parseCommand_exit() throws Exception {
+        assertTrue(parser.parse(ExitCommand.COMMAND_WORD) instanceof ExitCommand);
+        assertTrue(parser.parse(ExitCommand.COMMAND_WORD + " 3") instanceof ExitCommand);
+    }
+
+    @Test
+    public void parseCommand_help() throws Exception {
+        assertTrue(parser.parse(HelpCommand.COMMAND_WORD) instanceof HelpCommand);
+        assertTrue(parser.parse(HelpCommand.COMMAND_WORD + " 3") instanceof HelpCommand);
+    }
+}
